Refactor - move trap filter code into it's own file
