+++ Products/ZenEvents/tests/testTrapFilter.py
